SpiNorFlashDxe: Spansion S25FS512S workaround
The sector map table in Spansion S25FS512S is not giving the
correct Sector map configuration.

That is why this temp hack has been put in place.

DISCLAIMER : after using this hack we can not advertize
that this flash driver will work for all flash devices.
We can only advertixing the support for S25FS512S and MT35X
flash present in LS1046ARDB and LX2160ARDB respectivly.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>