Platform/RaspberryPi/RPi4: Fix non-standard ACPI HIDs
REF: https://github.com/pftf/RPi4/issues/67

Remove non-standard RPI ACPI _CIDs that are not needed. This
fixes and FWTS failure as well.

The windows drivers at https://github.com/raspberrypi/windows-drivers
should still be able to match the ACPI objects using the HIDs which
are supported in the drivers, with these two recent changes needed:
469702898789e555c6947e50216a3f79e0ddeb9
and
5c5e2742b4c983b3001c473b168b0dae2fcba0c2

Signed-off-by: Samer El-Haj-Mahmoud <Samer.El-Haj-Mahmoud@arm.com>